add toSoap for PointToPointService and QuerySummaryConfirmedRequest
